refactor: removed unused methods
