+++ scripts/deploy_data_models.py
@@ -1,4 +1,4 @@
-from cognite.powerops._cli import deploy
+from cognite.powerops.cli import deploy
 
 
 if __name__ == "__main__":
