+++ scripts/populate_data_models.py
@@ -1,4 +1,4 @@
-from cognite.powerops._cli import apply
+from cognite.powerops.cli import apply
 from tests.constants import REPO_ROOT
 
 
@@ -12,12 +12,7 @@ def main():
         "RKOMMarketDataModel",
     ]
     for model in models:
-        apply(
-            demo_data,
-            "dayahead",
-            [model],
-            auto_yes=True,
-        )
+        apply(demo_data, "dayahead", [model], auto_yes=True)
 
 
 if __name__ == "__main__":
